src/requirements.txt
aiohttp
discord.py
requests
replicate
